Simplify kill checks
Since it is now possible to be informed of all player or party kills via the
PARTY_KILL event, the checks on whether a kill should be credited or not
are no longer needed. All PARTY_KILLs are now counted, and kill counts are
no longer incremented when recording XP gain, to avoid double counting.

This changes the behaviour of the addon slightly: creatures which do not yield
XP are now always tracked when the player or a party kills them, not just when
the creature killed was the one last hit by the player.

More simply: if the game considers you eligible for kill credit, it's counted.

NB: This has not been tested when a party member kills something too far away
for the player to receive XP, but close enough to be recorded in the combat
log. If this situation can occur, then it will still be counted.