Add player death tracking
Now that UNIT_DIED correctly fires upon player death, reinstate the ablity
to track deaths. This should have the same behaviour as the feature
prior to 2.4.0 (ie. it guesses the killer based on who hit the player
last, according to the combat log. This means it may (in theory) suffer
from races, and can't track environmental deaths (falling, drowning,
lava, ...).)