Tighten test for when to try expiring creep data
When levelling up, check for low-level creep data to expire only if the
number of levels to store is less than the current character level.
This removes the last remaining use of the maxLevel variable, making it
redundant.
Additionally, this makes data expiration actually work when storeLevel >
70 but less than current character level, since this was hardcoded for
maxLevel in some places.